% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_all.R
\name{Add_all}
\alias{Add_all}
\title{Run all relevant scripts before you build the Morbid Genes Panel}
\usage{
Add_all(
  ClinVarCutoff = 4,
  HGMDCutoff = 4,
  .VarvisGeneManagement = VarvisGeneManagement,
  .morbidmap = morbidmap,
  .mim2gene = mim2gene,
 
    panelAppFile = "W:/HUG/04 Klinische Genomik/10 Panels/MorbidGenes-Panel/PanelAppGenes/2021_08_25_PA_all_genes.csv",
 
    manualFile = "W:/HUG/04 Klinische Genomik/10 Panels/MorbidGenes-Panel/GenesToBeAddedManually.xlsx",
  ...
)
}
\arguments{
\item{ClinVarCutoff}{ClinVar's pathogenic cutoff value. See \code{\link{Add_ClinVar}} for details}

\item{HGMDCutoff}{HGMD's pathogenic cutoff value. See \code{\link{Add_HGMD}} for details}

\item{.VarvisGeneManagement}{See \code{\link{Add_HGNC}} for details}

\item{.morbidmap}{See \code{\link{Add_HGNC}} for details}

\item{.mim2gene}{See \code{\link{Add_OMIM}} for details}

\item{...}{Additional arguments passed on to \code{\link{Add_ClinVar}}}

\item{PanelAppFile}{See \code{\link{Add_PanelApp}} for details}

\item{ManualFile}{See \code{\link{Add_ManualGenes}} for details}
}
\value{

}
\description{
This is a convenience wrapper script utilizing \code{\link{Add_ManualGenes}},
\code{\link{Add_PanelApp}}, \code{\link{Add_ClinVar}}, \code{\link{Add_OMIM}},
\code{\link{Add_HGMD}}, \code{\link{Add_HGNC}}.
}
\section{Warning}{

This script will add all the relevant variables to your global environment for convenience reasons
}

\examples{
Add_all(ClinVarCutoff = 4, mim2gene = mim2gene)
}
