% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_ClinVar.R
\name{Add_ClinVar}
\alias{Add_ClinVar}
\title{Add the data from both Clinvar files}
\usage{
Add_ClinVar(
  .clinvar_tsv_filtered = clinvar_tsv_filtered,
  .tidy_clinvar_vcf_meta = tidy_clinvar_vcf_meta,
  .mim2gene = mim2gene,
  cutoff = 4,
  print_NonMatching = T
)
}
\arguments{
\item{.clinvar_tsv_filtered}{ClinVar's filtered .tsv file.
Should be in your global environment after running \code{\link{StartNewVersion}}}

\item{.tidy_clinvar_vcf_meta}{ClinVar's filtered .vcf file.
Should be in your global environment after running \code{\link{StartNewVersion}}}

\item{.mim2gene}{mim2gene file from OMIM.
Should be in your global environment after running \code{\link{StartNewVersion}}}

\item{cutoff}{minimum number of pathogenic variants per gene to be included into the Morbid Genes panel}

\item{print_NonMatching}{Should the Transcriptgenes be printed which do not match the HGNC nomenclature?}
}
\value{
A dataframe with the ClinVar genes
}
\description{
Add the data from both Clinvar files
}
\examples{
\dontrun{
clinvar_tsv_filtered_patho_HGNC = Add_ClinVar(cutoff = 4)
}
}
