% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_GenCC.R
\name{Add_GenCC}
\alias{Add_GenCC}
\title{Add Genes from the GenCC Database}
\usage{
Add_GenCC(.gencc_tsv = gencc_tsv, classification = "Definitive")
}
\arguments{
\item{classification}{The confidence of GenCC-Genes. One of "Definitive" (default),
"Limited", "Moderate", "Refuted" or "All"}

\item{.gencc}{The GenCC tsv file.
Should be in your global environment after running \code{\link{StartNewVersion}}}
}
\description{
Add Genes from the GenCC Database
}
\examples{
\dontrun{
gencc = Add_GenCC()
}
}
