% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_HGMD.R
\name{Add_HGMD}
\alias{Add_HGMD}
\title{Add HGMD Data and determine pathogenic cutoff}
\usage{
Add_HGMD(cutoff = 4)
}
\arguments{
\item{cutoff}{minimum number of pathogenic variants per gene to be included into the Morbid Genes panel}
}
\value{
A dataframe with the HGMD genes
}
\description{
Add HGMD Data and determine pathogenic cutoff
}
\examples{
\dontrun{
HGMD_count_HGNC = Add_HGMD(cutoff = 4)
}
}
