% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_HGNC.R
\name{Add_HGNC}
\alias{Add_HGNC}
\title{Add the HGNC Data and genomic positions (GRCh38) to the Varvis Gene Management file}
\usage{
Add_HGNC(.VarvisGeneManagement = VarvisGeneManagement)
}
\arguments{
\item{.VarvisGeneManagement}{The Varvis Gene Management file.
Should be in your global environment after running \code{\link{StartNewVersion}}}
}
\value{

}
\description{
Add the HGNC Data and genomic positions (GRCh38) to the Varvis Gene Management file
}
\examples{
\dontrun{
VarvisGeneManagement_HGNC = Add_HGNC(.VarvisGeneManagement = VarvisGeneManagement)
}
}
