% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_ManualGenes.R
\name{Add_ManualGenes}
\alias{Add_ManualGenes}
\title{Add genes which have to be added manually from the Excel spreadsheet}
\usage{
Add_ManualGenes(
  ManualFile =
    "W:/HUG/04 Klinische Genomik/10 Panels/MorbidGenes-Panel/GenesToBeAddedManually.xlsx"
)
}
\arguments{
\item{ManualFile}{The path to the file where the Manual Genes are stored}
}
\value{
A dataframe with the manually added genes
}
\description{
Add genes which have to be added manually from the Excel spreadsheet
}
\examples{
\dontrun{
Manual = Add_ManualGenes()
}
}
