% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_OMIM.R
\name{Add_OMIM}
\alias{Add_OMIM}
\title{Reshape the morbidmap to be added to the Panel}
\usage{
Add_OMIM(.morbidmap = morbidmap, .mim2gene = mim2gene)
}
\arguments{
\item{.morbidmap}{The morbidmap file from OMIM.
Should be in your global environment after running \code{\link{StartNewVersion}}}

\item{.mim2gene}{The mim2gene file from OMIM.
Should be in your global environment after running \code{\link{StartNewVersion}}}
}
\value{
A dataframe with the OMIM data
}
\description{
Reshape the morbidmap to be added to the Panel
}
\examples{
\dontrun{
morbidmap_reshape = Add_OMIM(.morbidmap = morbidmap, .mim2gene = mim2gene)
}
}
