% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_PanelApp.R
\name{Add_PanelApp}
\alias{Add_PanelApp}
\title{Add the PanelApp Genes}
\usage{
Add_PanelApp(
 
    PanelAppFile = "W:/HUG/04 Klinische Genomik/10 Panels/MorbidGenes-Panel/PanelAppGenes/2021_08_25_PA_all_genes.csv"
)
}
\arguments{
\item{PanelAppFile}{The path to the file where the PanelApp Genes are stored}
}
\value{

}
\description{
Add the PanelApp Genes
}
\examples{
\dontrun{
PanelAppGenes = Add_PanelApp()
}
}
