% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_SysNDD.R
\name{Add_SysNDD}
\alias{Add_SysNDD}
\title{Add Genes from the SysNDD Database}
\usage{
Add_SysNDD(category = "Definitive", inheritance = "All", download = T)
}
\arguments{
\item{category}{The confidence of SysNDD-Genes. One of "Definitive" (default),
"Limited", "Moderate", "Refuted" or "All"}

\item{inheritance}{Filter for specific inheritance patterns. One of "All" (default),
"Dominant", "Other", "Recessive" or "X-linked"}

\item{download}{Should the latest SysNDD be downloaded? If not, previous version will be used}
}
\value{
A dataframe with the SysNDD genes
}
\description{
Add Genes from the SysNDD Database
}
\examples{
\dontrun{
SysNDD = Add_SysNDD()
}
}
