% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_MorbidGenesPanel.R
\name{Build_MorbidGenesPanel}
\alias{Build_MorbidGenesPanel}
\title{Title}
\usage{
Build_MorbidGenesPanel(
  directory = "W:/HUG/04 Klinische Genomik/10 Panels/MorbidGenes-Panel/",
  version = format(Sys.time(), "\%Y-\%m"),
  .VarvisGeneManagement_HGNC = VarvisGeneManagement_HGNC,
  .clinvar_tsv_filtered_patho_HGNC = clinvar_tsv_filtered_patho_HGNC,
  .morbidmap_reshape = morbidmap_reshape,
  .Manual = Manual,
  .PanelAppGenes = PanelAppGenes,
  save = T
)
}
\arguments{
\item{directory}{The same you provided in \code{\link{StartNewVersion}}}

\item{version}{The same you provided in \code{\link{StartNewVersion}}}

\item{.VarvisGeneManagement_HGNC}{The result of the function \code{\link{Add_HGNC}}
or in your Global Env after running \code{\link{Add_all}}}

\item{.clinvar_tsv_filtered_patho_HGNC}{The result of the function \code{\link{Add_ClinVar}}
or in your Global Env after running \code{\link{Add_all}}}

\item{.morbidmap_reshape}{The result of the function \code{\link{Add_OMIM}}
or in your Global Env after running \code{\link{Add_all}}}

\item{.Manual}{The result of the function \code{\link{Add_ManualGenes}}
or in your Global Env after running \code{\link{Add_all}}}

\item{.PanelAppGenes}{The result of the function \code{\link{Add_PanelApp}}
or in your Global Env after running \code{\link{Add_all}}}

\item{save}{should the Panel be saved as a .csv? Defaults to Yes in the corresponding directory}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
Build_MorbidGenesPanel()
}
}
