% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Morbidfiles.R
\name{Load_Morbidfiles}
\alias{Load_Morbidfiles}
\title{Load all relevant files after building the Morbid Genes Panel in
case they are not in your environment anymore}
\usage{
Load_Morbidfiles(directory, .ClinVarCutoff = 4, .HGMDCutoff = 4)
}
\arguments{
\item{directory}{The directory where all the files are saved, e.g. "MorbidGenes-Panel-v2021-12.1}

\item{.ClinVarCutoff}{ClinVar's pathogenic cutoff value. See \code{\link{Add_ClinVar}} for details}

\item{.HGMDCutoff}{HGMD's pathogenic cutoff value. See \code{\link{Add_HGMD}} for details}
}
\value{
Loads all files to the global environment
}
\description{
This is a convenience wrapper script utilizing \code{\link{Add_ManualGenes}},
\code{\link{Add_PanelApp}}, \code{\link{Add_ClinVar}}, \code{\link{Add_OMIM}},
\code{\link{Add_HGMD}}, \code{\link{Add_HGNC}}.
}
\section{Warning}{

This script will add all the relevant variables to your global environment for convenience reasons
}

\examples{
\dontrun{
Load_Morbidfiles(directory = "path/to/morbidgenesfiles")
}
}
