% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartNewVersion.R
\name{StartNewVersion}
\alias{StartNewVersion}
\title{Start New Morbid Genes Panel Version}
\usage{
StartNewVersion(
  directory = "W:/HUG/04 Klinische Genomik/10 Panels/MorbidGenes-Panel/",
  version = format(Sys.time(), "\%Y-\%m"),
  download_varvis = T,
  download_HGNC = T,
  download_OMIM = T,
  download_ClinVar_tsv = T,
  download_ClinVar_vcf = T,
  download_GenCC = T
)
}
\arguments{
\item{directory}{Where the files will be saved. Defaults to the Working Group directory where the other panel versions are saved}

\item{version}{The panel version. Defaults to "Year-Month"}

\item{download_varvis}{Should the Varvis Gene Management table be downloaded?}

\item{download_HGNC}{Should the current HGNC nomenclature be downloaded?}

\item{download_OMIM}{Should the current OMIM files be downloaded?}

\item{download_ClinVar_tsv}{Should the current ClinVar tsv file be downloaded?}

\item{download_ClinVar_vcf}{Should the current ClinVar vcf file be downloaded?}

\item{download_GenCC}{Should the current ClinVar vcf file be downloaded?}
}
\value{
Generates a new folder and downloads data
}
\description{
Start New Morbid Genes Panel Version
}
\details{
The files needed for the next step will also be added to the global environment.
If you do not download the latest datasets, the ones from the previous version
will be used.

It's best to stick with the default values, otherwise things will get
confusing. This script will also automatically add a subversion in case there
is a previous version from the same month, to avoid overwriting.
}
\examples{
\dontrun{
StartNewVersion(download_varvis = F, download_HGNC = F,
download_OMIM = F, download_ClinVar_tsv = F, download_ClinVar_vcf = F)
# This will create a panel directory in the default HUG panel folder,
# the version will be "Year-Month" and a subversion
}

}
