% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_all.R
\name{Add_all}
\alias{Add_all}
\title{Run all relevant scripts before you build the Morbid Genes Panel}
\usage{
Add_all(
  ClinVarCutoff = 4,
  HGMDCutoff = 4,
  .morbidmap = morbidmap,
  .mim2gene = mim2gene,
  PanelApp_Panel = "both",
  PanelApp_confidence = 3,
  downloadPanelApp = T,
  SysNDD_category = "Definitive",
  SysNDD_inheritance = "All",
  download_SysNDD = T,
  .gencc_tsv = gencc_tsv,
  gencc_classification = "Definitive",
  manualFile =
    "W:/HUG/04 Klinische Genomik/10 Panels/MorbidGenes-Panel/GenesToBeAddedManually.xlsx",
  ...
)
}
\arguments{
\item{ClinVarCutoff}{ClinVar's pathogenic cutoff value. See \code{\link{Add_ClinVar}} for details}

\item{HGMDCutoff}{HGMD's pathogenic cutoff value. See \code{\link{Add_HGMD}} for details}

\item{.morbidmap}{See \code{\link{Add_HGNC}} for details}

\item{.mim2gene}{See \code{\link{Add_OMIM}} for details}

\item{PanelApp_Panel}{See \code{\link{Add_PanelApp}} for details}

\item{PanelApp_confidence}{See \code{\link{Add_PanelApp}} for details}

\item{downloadPanelApp}{See \code{\link{Add_PanelApp}} for details}

\item{SysNDD_category}{See \code{\link{Add_SysNDD}} for details}

\item{SysNDD_inheritance}{See \code{\link{Add_SysNDD}} for details}

\item{download_SysNDD}{See \code{\link{Add_SysNDD}} for details}

\item{.gencc_tsv}{See \code{\link{Add_GenCC}} for details}

\item{gencc_classification}{See \code{\link{Add_GenCC}} for details}

\item{manualFile}{See \code{\link{Add_ManualGenes}} for details}

\item{...}{Additional arguments passed on to \code{\link{Add_ClinVar}}}
}
\description{
This is a convenience wrapper script utilizing \code{\link{Add_ManualGenes}},
\code{\link{Add_PanelApp}}, \code{\link{Add_ClinVar}}, \code{\link{Add_OMIM}},
\code{\link{Add_HGMD}}, \code{\link{Add_HGNC}}.
}
\section{Warning}{

This script will add all the relevant variables to your global environment for convenience reasons
}

\examples{
\dontrun{
Add_all(ClinVarCutoff = 4, mim2gene = mim2gene)
}
}
