% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_HGNC.R
\name{Add_HGNC}
\alias{Add_HGNC}
\title{Add the HGNC Data and genomic positions (GRCh38) to the Varvis Gene Management file}
\usage{
Add_HGNC(
  .VarvisGeneManagement = VarvisGeneManagement,
  directory = "W:/HUG/04 Klinische Genomik/10 Panels/MorbidGenes-Panel/",
  version = format(Sys.time(), "\%Y-\%m"),
  add_coordinates = T
)
}
\arguments{
\item{.VarvisGeneManagement}{The Varvis Gene Management file.
Should be in your global environment after running \code{\link{StartNewVersion}}}

\item{directory}{The same you provided in \code{\link{StartNewVersion}}}

\item{version}{The same you provided in \code{\link{StartNewVersion}}}

\item{add_coordinates}{Add genomic coordinates from Ensembl. Can be set to F if Ensembl is down}
}
\value{
A dataframe with the HGNC Data
}
\description{
Add the HGNC Data and genomic positions (GRCh38) to the Varvis Gene Management file
}
\examples{
\dontrun{
VarvisGeneManagement_HGNC = Add_HGNC(.VarvisGeneManagement = VarvisGeneManagement)
}
}
