% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_PanelApp.R
\name{Add_PanelApp}
\alias{Add_PanelApp}
\title{Add the PanelApp Genes}
\usage{
Add_PanelApp(Panel = "both", confidence = 3, download = T)
}
\arguments{
\item{Panel}{One of "UK", "AUS" or "both" (default). This is to specify which PanelApp to use}

\item{confidence}{The minimum confidence of genes to keep. Can be 1 (minimum red genes),
2 (minimum amber genes) or the default 3 (minimum green genes)}

\item{download}{Should the latest data be downloaded? If not, the previous versions will be read}
}
\value{
A dataframe with the PanelApp genes
}
\description{
Add the PanelApp Genes
}
\examples{
\dontrun{
PanelAppGenes = Add_PanelApp()
}
}
